+++ Animal.java
@@ -1,5 +1,7 @@
 
+import java.util.Iterator;
 import java.util.List;
+import java.util.Random;
 
 /**
  * A class representing shared characteristics of animals.
@@ -15,6 +17,23 @@ public abstract class Animal
     private Field field;
     // The animal's position in the field.
     private Location location;
+
+    private boolean isMale;
+
+    private Random rand = Randomizer.getRandom();
+
+    // The age at which a fox can start to breed.
+    protected static int BREEDING_AGE;
+    // The age to which a fox can live.
+    protected static int MAX_AGE;
+    // The likelihood of a fox breeding.
+    protected static double BREEDING_PROBABILITY;
+    // The maximum number of births.
+    protected static int MAX_LITTER_SIZE;
+    // The food value of a single rabbit. In
+
+    protected int age;
+    protected int foodLevel;
     
     /**
      * Create a new animal at location in field.
@@ -27,14 +46,115 @@ public abstract class Animal
         alive = true;
         this.field = field;
         setLocation(location);
+        decideGender();
     }
     
+    private void decideGender(){
+        isMale = rand.nextBoolean();
+    }
+
     /**
      * Make this animal act - that is: make it do
      * whatever it wants/needs to do.
      * @param newAnimals A list to receive newly born animals.
      */
-    abstract public void act(List<Animal> newAnimals);
+    public void act(List<Animal> newAnimals) {
+        incrementAge();
+        incrementHunger();
+        if(isAlive()) {
+                      
+            // Move towards a source of food if found.
+            Location newLocation = findFood();
+            if(newLocation == null) { 
+                // No food found - try to move to a free location.
+                newLocation = findMate();
+                if(newLocation == null) {
+                    newLocation = getField().freeAdjacentLocation(getLocation(), true);
+                } else {
+                    giveBirth(newAnimals);
+                }
+                
+            } 
+            // See if it was possible to move.
+            if(newLocation != null) {
+                setLocation(newLocation);
+            }
+            else {
+                // Overcrowding.
+                setDead();
+            }
+        }
+    };
+    
+    protected abstract Location findFood();
+
+    /**
+     * Increase the age. This could result in the fox's death.
+     */
+    private void incrementAge()
+    {
+        age++;
+        if(age > MAX_AGE) {
+            //setDead();
+        }
+    }
+    
+/**
+     * Check whether or not this fox is to give birth at this step.
+     * New births will be made into free adjacent locations.
+     * @param newPredators A list to return newly born foxes.
+     */
+    private void giveBirth(List<Animal> newPredators)
+    {
+        // New foxes are born into adjacent locations.
+        // Get a list of adjacent free locations.
+        Field field = getField();
+        List<Location> free = field.getFreeAdjacentLocations(getLocation(), true);
+        int births = breed();
+        for(int b = 0; b < births && free.size() > 0; b++) {
+            Location loc = free.remove(0);
+            Animal young = copyThis(loc);
+            newPredators.add(young);
+        }
+    }
+
+    /**
+     * Provides a function for a subclass to create a copy of itself
+     */
+    protected abstract Animal copyThis(Location loc);
+
+    /**
+     * Generate a number representing the number of births,
+     * if it can breed.
+     * @return The number of births (may be zero).
+     */
+    private int breed()
+    {
+        int births = 0;
+        if(canBreed() && rand.nextDouble() <= BREEDING_PROBABILITY) {
+            births = rand.nextInt(MAX_LITTER_SIZE) + 1;
+        }
+        return births;
+    }
+
+    /**
+     * A fox can breed if it has reached the breeding age.
+     */
+    private boolean canBreed()
+    {
+        return age >= BREEDING_AGE;
+    }
+
+    /**
+     * Make this fox more hungry. This could result in the fox's death.
+     */
+    private void incrementHunger()
+    {
+        foodLevel--;
+        if(foodLevel <= 0) {
+            setDead();
+        }
+    }
 
     /**
      * Check whether the animal is alive or not.
@@ -43,6 +163,11 @@ public abstract class Animal
     protected boolean isAlive()
     {
         return alive;
+    }
+
+    public boolean isOppositeGender(boolean gender)
+    {
+        return gender ^ isMale;
     }
 
     /**
@@ -68,6 +193,28 @@ public abstract class Animal
         return location;
     }
     
+    protected Location findMate()
+    {
+        Field field = getField();
+        List<Location> adjacent = field.adjacentLocations(getLocation());
+        Iterator<Location> it = adjacent.iterator();
+        while(it.hasNext()) {
+            Location where = it.next();
+            Object object = field.getObjectAt(where);
+            Animal animal;
+            if(object instanceof Animal){
+                animal = (Animal) object;
+                if(animal.getClass().equals(this.getClass()) && animal.isOppositeGender(isMale)) {
+                    if(animal.isAlive()) { 
+                        return where;
+                    }
+                }
+            }
+        }
+        return null;
+    }
+
+
     /**
      * Place the animal at the new location in the given field.
      * @param newLocation The animal's new location.
