+++ Predator.java
@@ -6,25 +6,11 @@ public abstract class Predator extends Animal{
 
 // Characteristics shared by all foxes (class variables).
     
-    // The age at which a fox can start to breed.
-    private static int BREEDING_AGE;
-    // The age to which a fox can live.
-    private static int MAX_AGE;
-    // The likelihood of a fox breeding.
-    private static double BREEDING_PROBABILITY;
-    // The maximum number of births.
-    private static int MAX_LITTER_SIZE;
     // The food value of a single rabbit. In effect, this is the
     // number of steps a fox can go before it has to eat again.
     private static int PREY_FOOD_VALUE;
     // A shared random number generator to control breeding.
     private Random rand = Randomizer.getRandom();
-
-    // Individual characteristics (instance fields).
-    // The fox's age.
-    protected int age;
-    // The fox's food level, which is increased by eating rabbits.
-    protected int foodLevel;
 
     public Predator(boolean randomAge, Field field, Location location, int breedingAge, int maxAge, double breedingProb, int maxLitterSize, int preyFoodValue) {
         super(field, location);
@@ -37,7 +23,7 @@ public abstract class Predator extends Animal{
 
         if(randomAge) {
             age = rand.nextInt(MAX_AGE);
-            foodLevel = rand.nextInt(PREY_FOOD_VALUE);
+            foodLevel = PREY_FOOD_VALUE;
         }
         else {
             age = 0;
@@ -45,126 +31,20 @@ public abstract class Predator extends Animal{
         }
     }
     
-    /**
-     * This is what the fox does most of the time: it hunts for
-     * rabbits. In the process, it might breed, die of hunger,
-     * or die of old age.
-     * @param field The field currently occupied.
-     * @param newPredators A list to return newly born foxes.
-     */
-    public void act(List<Animal> newPredators)
-    {
-        incrementAge();
-        incrementHunger();
-        if(isAlive()) {
-            giveBirth(newPredators);            
-            // Move towards a source of food if found.
-            Location newLocation = findFood();
-            if(newLocation == null) { 
-                // No food found - try to move to a free location.
-                newLocation = getField().freeAdjacentLocation(getLocation(), true);
-            }
-            // See if it was possible to move.
-            if(newLocation != null) {
-                setLocation(newLocation);
-            }
-            else {
-                // Overcrowding.
-                setDead();
-            }
-        }
-    }
-
-    /**
-     * Increase the age. This could result in the fox's death.
-     */
-    private void incrementAge()
-    {
-        age++;
-        if(age > MAX_AGE) {
-            setDead();
-        }
-    }
-    
-    /**
-     * Make this fox more hungry. This could result in the fox's death.
-     */
-    private void incrementHunger()
-    {
-        foodLevel--;
-        if(foodLevel <= 0) {
-            setDead();
-        }
-    }
     
     /**
      * Look for rabbits adjacent to the current location.
      * Only the first live rabbit is eaten.
      * @return Where food was found, or null if it wasn't.
      */
-    protected Location findFood()
-    {
-        Field field = getField();
-        List<Location> adjacent = field.adjacentLocations(getLocation());
-        Iterator<Location> it = adjacent.iterator();
-        while(it.hasNext()) {
-            Location where = it.next();
-            Object animal = field.getObjectAt(where);
-            if(animal instanceof Prey) {
-                Prey rabbit = (Prey) animal;
-                if(rabbit.isAlive()) { 
-                    rabbit.setDead();
-                    foodLevel = PREY_FOOD_VALUE;
-                    return where;
-                }
-            }
-        }
-        return null;
-    }
-    
-    /**
-     * Check whether or not this fox is to give birth at this step.
-     * New births will be made into free adjacent locations.
-     * @param newPredators A list to return newly born foxes.
-     */
-    private void giveBirth(List<Animal> newPredators)
-    {
-        // New foxes are born into adjacent locations.
-        // Get a list of adjacent free locations.
-        Field field = getField();
-        List<Location> free = field.getFreeAdjacentLocations(getLocation(), true);
-        int births = breed();
-        for(int b = 0; b < births && free.size() > 0; b++) {
-            Location loc = free.remove(0);
-            Predator young = copyThis(loc);
-            newPredators.add(young);
-        }
-    }
+    @Override
+    protected abstract Location findFood();
     
     /**
      * Provides a function for a subclass to create a copy of itself
      */
+    @Override
     protected abstract Predator copyThis(Location loc);
 
-    /**
-     * Generate a number representing the number of births,
-     * if it can breed.
-     * @return The number of births (may be zero).
-     */
-    private int breed()
-    {
-        int births = 0;
-        if(canBreed() && rand.nextDouble() <= BREEDING_PROBABILITY) {
-            births = rand.nextInt(MAX_LITTER_SIZE) + 1;
-        }
-        return births;
-    }
-
-    /**
-     * A fox can breed if it has reached the breeding age.
-     */
-    private boolean canBreed()
-    {
-        return age >= BREEDING_AGE;
-    }
+    
 }
