+++ Prey.java
@@ -5,14 +5,7 @@ import java.util.Random;
 public abstract class Prey extends Animal {
    // Characteristics shared by all rabbits (class variables).
 
-    // The age at which a rabbit can start to breed.
-    private static int BREEDING_AGE;
-    // The age to which a rabbit can live.
-    private static int MAX_AGE;
-    // The likelihood of a rabbit breeding.
-    private static double BREEDING_PROBABILITY;
-    // The maximum number of births.
-    private static int MAX_LITTER_SIZE;
+    
     // The food value of a single rabbit. In effect, this is the
     // number of steps a fox can go before it has to eat again.
     private static int PLANT_FOOD_VALUE = 20;
@@ -21,8 +14,6 @@ public abstract class Prey extends Animal {
     
     // Individual characteristics (instance fields).
     
-    // The rabbit's age.
-    private int age;
 // The fox's food level, which is increased by eating rabbits.
 protected int foodLevel;
     /**
@@ -44,92 +35,24 @@ protected int foodLevel;
         
         if(randomAge) {
             age = rand.nextInt(MAX_AGE);
-            foodLevel = rand.nextInt(PLANT_FOOD_VALUE);
+            foodLevel = PLANT_FOOD_VALUE;
         }
         else {
             age = 0;
             foodLevel = PLANT_FOOD_VALUE;
         }
     }
-    
-    /**
-     * This is what the rabbit does most of the time - it runs 
-     * around. Sometimes it will breed or die of old age.
-     * @param newPrey A list to return newly born rabbits.
-     */
-    public void act(List<Animal> newPrey)
-    {
-        incrementAge();
-        incrementHunger();
-        if(isAlive()) {
-            giveBirth(newPrey);            
-            // Move towards a source of food if found.
-            Location newLocation = findFood();
-            if(newLocation == null) { 
-                // No food found - try to move to a free location.
-                newLocation = getField().freeAdjacentLocation(getLocation(), true);
-            }
-            // See if it was possible to move.
-            if(newLocation != null) {
-                setLocation(newLocation);
-            }
-            else {
-                // Overcrowding.
-                setDead();
-            }
-        }
-    }
 
-    /**
-     * Increase the age.
-     * This could result in the rabbit's death.
-     */
-    private void incrementAge()
-    {
-        age++;
-        if(age > MAX_AGE) {
-            setDead();
-        }
-    }
-    
-    /**
-     * Check whether or not this rabbit is to give birth at this step.
-     * New births will be made into free adjacent locations.
-     * @param newPrey A list to return newly born rabbits.
-     */
-    private void giveBirth(List<Animal> newPrey)
-    {
-        // New rabbits are born into adjacent locations.
-        // Get a list of adjacent free locations.
-        Field field = getField();
-        List<Location> free = field.getFreeAdjacentLocations(getLocation(), true);
-        int births = breed();
-        for(int b = 0; b < births && free.size() > 0; b++) {
-            Location loc = free.remove(0);
-            Prey young = copyThis(loc);
-            newPrey.add(young);
-        }
-    }
-
+    @Override
     protected abstract Prey copyThis(Location loc);
         
-
-    /**
-     * Make this fox more hungry. This could result in the fox's death.
-     */
-    private void incrementHunger()
-    {
-        foodLevel--;
-        if(foodLevel <= 0) {
-            setDead();
-        }
-    }
     
     /**
      * Look for rabbits adjacent to the current location.
      * Only the first live rabbit is eaten.
      * @return Where food was found, or null if it wasn't.
      */
+    @Override
     protected Location findFood()
     {
         Field field = getField();
@@ -148,28 +71,5 @@ protected int foodLevel;
             }
         }
         return null;
-    }
-
-    /**
-     * Generate a number representing the number of births,
-     * if it can breed.
-     * @return The number of births (may be zero).
-     */
-    private int breed()
-    {
-        int births = 0;
-        if(canBreed() && rand.nextDouble() <= BREEDING_PROBABILITY) {
-            births = rand.nextInt(MAX_LITTER_SIZE) + 1;
-        }
-        return births;
-    }
-
-    /**
-     * A rabbit can breed if it has reached the breeding age.
-     * @return true if the rabbit can breed, false otherwise.
-     */
-    private boolean canBreed()
-    {
-        return age >= BREEDING_AGE;
     }
 }
